This is task that is assigned by the GARUNA company as a test task for my interview
Here my task was to fetch the data from the API and display it on the UNITY UI

I have tried to create this applicaiton suitable for android and windows both

but may it will not be able to work on android because it need to be connected to one server.

But in windows it could run as it could run in local server using nodejs.


I took 3 hours to complete this task as for the first time I was working on node js.

Anyway I learnt so many things.

Thank You,
Regards,
Vedprakash Saidarsan Sahoo